#### :question: What

Add `Copy` method to the `Record` interface.
This is necessary for filters that want to create copies of a record.
Without that new method, one would have to call `src.ToText` followed by `dst.Parse` to make `dst` being a copy of `src` `Record`. This is what I call the naive solution.

The following benchmarks comparisons shows the optimized Copy vs the naive version:

    name                  old time/op    new time/op    delta
    LogLineCopy/set=0-8     7.94µs ± 8%    4.65µs ±10%  -41.43%  (p=0.000 n=15+15)
    LogLineCopy/set=1-8     32.7µs ± 5%    30.8µs ± 6%   -5.86%  (p=0.000 n=13+15)
    LogLineCopy/set=10-8    33.8µs ± 3%    30.6µs ± 5%   -9.56%  (p=0.000 n=14+15)
    LogLineCopy/set=20-8    32.8µs ± 6%    30.1µs ± 4%   -8.03%  (p=0.000 n=15+15)

    name                  old alloc/op   new alloc/op   delta
    LogLineCopy/set=0-8     24.5kB ± 0%    24.5kB ± 0%     ~     (all equal)
    LogLineCopy/set=1-8     39.0kB ± 0%    31.3kB ± 0%  -19.71%  (p=0.000 n=15+15)
    LogLineCopy/set=10-8    39.0kB ± 0%    31.3kB ± 0%  -19.71%  (p=0.000 n=15+15)
    LogLineCopy/set=20-8    39.3kB ± 0%    31.3kB ± 0%  -20.50%  (p=0.000 n=15+15)

    name                  old allocs/op  new allocs/op  delta
    LogLineCopy/set=0-8       3.00 ± 0%      3.00 ± 0%     ~     (all equal)
    LogLineCopy/set=1-8       6.00 ± 0%      4.00 ± 0%  -33.33%  (p=0.000 n=15+15)
    LogLineCopy/set=10-8      6.00 ± 0%      4.00 ± 0%  -33.33%  (p=0.000 n=15+15)
    LogLineCopy/set=20-8      6.00 ± 0%      4.00 ± 0%  -33.33%  (p=0.000 n=15+15)


#### :hammer: How to test

1. List all steps necessary;
2. To test this pull request.

#### :white_check_mark: Checklists

_This section contains a list of checklists for common uses, please delete the checklists that are useless for your current use case (or add another checklist if your use case isn't covered yet)._

- [x] Is there unit/integration test coverage for all new and/or changed functionality added in this PR?
- [ ] Have the changes in this PR been functionally tested?
- [x] Has `make gofmt-write` been run on the code?
- [x] Has `make govet` been run on the code? Has the code been fixed accordingly to the output?
- [x] Have the changes been added to the [CHANGELOG.md](./CHANGELOG.md) file?
- [x] Have the steps in [CONTRIBUTING.md](./CONTRIBUTING.md) been followed to update a Go module?